requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
pyzbar==0.1.9
pdf2image==1.16.3
Pillow==10.0.1
pillow-heif==0.13.0
python-multipart==0.0.6
werkzeug==2.3.7
numpy<2.0
pytesseract==0.3.10
python-dotenv==1.0.0
